requirements.txt
motor==2.3.0
pymongo==3.11.0
PyYAML==5.3.1
six==1.15.0
playsound==1.2.2
